finish modular_network.from_args, dim to class attribute
