+++ adept/agents/impala/actor_critic_vtrace.py
@@ -20,7 +20,7 @@ from torch.nn import functional as F
 from adept.environments.env_registry import Engines
 from adept.expcaches.rollout import RolloutCache
 from adept.utils.util import listd_to_dlist, dlist_to_listd
-from adept.networks._base import ModularNetwork
+from adept.networks.modular_network import ModularNetwork
 from adept.agents.agent_module import AgentModule
