+++ adept/networks/_base.py
@@ -16,7 +16,6 @@ import abc
 
 import torch
 from adept.utils.requires_args import RequiresArgs
-from collections import ChainMap
 
 
 class BaseNetwork(torch.nn.Module):
@@ -64,204 +63,6 @@ class NetworkHead(torch.nn.Module):
         }, internals
 
 
-class ModularNetwork(BaseNetwork, metaclass=abc.ABCMeta):
-    def __init__(
-        self,
-        obs_key_to_submod,
-        body_submodule,
-        head_submodules,
-        output_space
-    ):
-        """
-        :param obs_key_to_submod: Dict[ObsKey, SubModule]
-        :param body_submodule: SubModule
-        :param head_submodules: List[SubModule]
-        :param output_space: Dict[OutputKey, Shape]
-        """
-        # modular network doesn't have networks for unused inputs
-        super().__init__()
-
-        # Input Nets
-        self._obs_keys = list(obs_key_to_submod.keys())
-        self.obs_key_to_input_submod = torch.nn.ModuleDict(
-            [(key, net) for key, net in obs_key_to_submod.items()]
-        )
-
-        # Body
-        self.body_submodule = body_submodule
-
-        # Heads
-        self.dim_to_head = torch.nn.ModuleDict(
-            [(submod.dim, submod) for submod in head_submodules]
-        )
-
-        # Outputs
-        self._output_keys = list(output_space.keys())
-        self.output_space = output_space
-        outputs = []
-        for output_name, shape in self.output_space.items():
-            dim = len(shape)
-            if dim == 1:
-                layer = torch.nn.Linear(
-                    self.head_submodules[dim].output_shape()[0], shape[0]
-                )
-            elif dim == 2:
-                layer = torch.nn.Conv1d(
-                    self.head_submodules[dim].output_shape()[0], shape[0],
-                    kernel_size=1
-                )
-            elif dim == 3:
-                layer = torch.nn.Conv2d(
-                    self.head_submodules[dim].output_shape()[0], shape[0],
-                    kernel_size=1
-                )
-            elif dim == 4:
-                layer = torch.nn.Conv3d(
-                    self.head_submodules[dim].output_shape()[0], shape[0],
-                    kernel_size=1
-                )
-            else:
-                raise ValueError('Invalid dim {}'.format(dim))
-            outputs.append((output_name, layer))
-        self.output_key_to_layer = torch.nn.ModuleDict(outputs)
-        self._validate_shapes()
-
-    def _validate_shapes(self):
-        # heads can't be higher dim than body
-        assert all([
-            head.dim <= self.body_submodule.dim
-            for head in self.head_submodules
-        ])
-        # output dims must have a corresponding head of the same dim
-        pass
-        # non feature dims of input submodules must be same
-        pass
-        # non-feature dimensions of heads must match desired output_shape
-        pass
-
-    @classmethod
-    def from_args(
-        cls,
-        args,
-        observation_space,
-        output_space,
-        network_registry
-    ):
-        """
-        Construct a Modular Network from arguments.
-
-        :param args: Dict[ArgName, Any]
-        :param observation_space: Dict[ObsKey, Shape]
-        :param output_space: Dict[OutputKey, Shape]
-        :param network_registry: NetworkRegistry
-        :return: ModularNetwork
-        """
-        # SubModule ID is ObsKey
-        id = 0
-        # Dict[ObsKey, SubModule]
-        # for shape in observation space, get dim
-        # instantiate input submodule of that dim
-        obs_key_to_submod = {}
-        for obs_key, shape in observation_space.items():
-            dim = len(shape)
-            if dim == 1:
-                submod = network_registry.lookup_submodule(
-                    args.net1d).from_args(args, shape, obs_key)
-            elif dim == 2:
-                submod = network_registry.lookup_submodule(
-                    args.net2d).from_args(args, shape, obs_key)
-            elif dim == 3:
-                submod = network_registry.lookup_submodule(
-                    args.net3d).from_args(args, shape, obs_key)
-            elif dim == 4:
-                submod = network_registry.lookup_submodule(
-                    args.net4d).from_args(args, shape, obs_key)
-            else:
-                raise ValueError('Invalid dim: {}'.format(dim))
-            obs_key_to_submod[obs_key] = submod
-
-        # SubModule
-        input_shape = None  # TODO
-        body_submod = network_registry.lookup_submodule(args.netbody).from_args(
-            args, input_shape, 'body')
-
-        # List[SubModule]
-        # instantiate heads based on output_shapes
-        head_submodules = []
-        for output_key, shape in output_space.items():
-            dim = len(shape)
-            if dim == 1:
-                submod = network_registry.lookup_submodule(
-                    args.head1d).from_args(args, shape, output_key)
-            elif dim == 2:
-                submod = network_registry.lookup_submodule(
-                    args.head2d).from_args(args, shape, output_key)
-            elif dim == 3:
-                submod = network_registry.lookup_submodule(
-                    args.head3d).from_args(args, shape, output_key)
-            elif dim == 4:
-                submod = network_registry.lookup_submodule(
-                    args.head4d).from_args(args, shape, output_key)
-            else:
-                raise ValueError('Invalid dim: {}'.format(dim))
-            head_submodules.append(submod)
-        return cls(
-            obs_key_to_submod, body_submod, head_submodules, output_space
-        )
-
-    def forward(self, obs_key_to_obs, internals):
-        """
-
-        :param obs_key_to_obs: Dict[str, torch.Tensor (1D | 2D | 3D | 4D)]
-        :param internals: Dict[str, torch.Tensor (ND)]
-        :return: Tuple[
-            Dict[str, torch.Tensor (1D | 2D | 3D | 4D)],
-            ChainMap[str, torch.Tensor (ND)]
-        ]
-        """
-        # Process input networks
-        nxt_internals = []
-        processed_inputs = []
-        for key in self._obs_keys:
-            result, nxt_internal = self.obs_key_to_input_submod[key].forward(
-                obs_key_to_obs[key],
-                internals,
-                dim=self.body_submodule.dim
-            )
-            processed_inputs.append(result)
-            nxt_internals.append(nxt_internal)
-
-        # Process body
-        body_out, nxt_internal = self.body.forward(
-            torch.cat(processed_inputs, dim=1),
-            internals
-        )
-
-        # Process heads
-        head_dim_to_head_out = {}
-        for head_submod in self.head_submodules:
-            head_out, nxt_internal = head_submod.forward(
-                self.body_submodule.to_dim(body_out, head_submod.dim),
-                internals
-            )
-            head_dim_to_head_out[head_submod.dim] = head_out
-            nxt_internals.append(nxt_internal)
-
-        # Process final outputs
-        output_key_to_output = {}
-        for key in self._output_keys:
-            output = self.output_key_to_layer[key](
-                head_dim_to_head_out[len(self.output_space[key])]
-            )
-            output_key_to_output[key] = output
-
-        return output_key_to_output, ChainMap(*internals)
-
-    def new_internals(self, device):
-        # TODO
-        return self.body.new_internals(device)
-
-
 class NetworkBody(torch.nn.Module, RequiresArgs, metaclass=abc.ABCMeta):
     @classmethod
     @abc.abstractmethod
@@ -299,91 +100,3 @@ class NetworkJunction(torch.nn.Module):
         for name, input_net in self.input_nets.named_children():
             embeddings.append(input_net.forward(obs_dict[name]))
         return torch.cat(embeddings, dim=1)
-
-
-class SubModule(torch.nn.Module, RequiresArgs, metaclass=abc.ABCMeta):
-    """
-    SubModule of a ModularNetwork.
-    """
-    def __init__(self, input_shape):
-        super(SubModule, self).__init__()
-        self._input_shape = input_shape
-
-    @classmethod
-    @abc.abstractmethod
-    def from_args(cls, args, input_shape):
-        raise NotImplementedError
-
-    @property
-    @abc.abstractmethod
-    def _output_shape(self):
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def output_shape(self, dim=None):
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def _forward(self, *input):
-        """
-        :param input: torch.Tensor (1D | 2D | 3D | 4D)
-        :return: Tuple[Result, Internals]
-        """
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def _to_1d(self, submodule_output):
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def _to_2d(self, submodule_output):
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def _to_3d(self, submodule_output):
-        raise NotImplementedError
-
-    @abc.abstractmethod
-    def _to_4d(self, submodule_output):
-        raise NotImplementedError
-
-    @property
-    def dim(self):
-        return len(self._input_shape)
-
-    @property
-    def input_shape(self):
-        return self._input_shape
-
-    def to_dim(self, submodule_output, dim):
-        """
-        :param submodule_output: torch.Tensor (1D | 2D | 3D | 4D)
-        Output of a forward pass to be converted.
-        :param dim: int Desired dimensionality
-        :return:
-        """
-        if dim <= 0 or dim > 4:
-            raise ValueError('Invalid dim: {}'.format(dim))
-        elif dim == 1:
-            return self._to_1d(submodule_output)
-        elif dim == 2:
-            return self._to_2d(submodule_output)
-        elif dim == 3:
-            return self._to_3d(submodule_output)
-        elif dim == 4:
-            return self._to_4d(submodule_output)
-
-    def forward(self, *input, dim=None):
-        submodule_output, internals = self._forward(*input)
-        if dim is None:
-            return submodule_output, internals
-        if dim == 1:
-            return self._to_1d(submodule_output), internals
-        elif dim == 2:
-            return self._to_2d(submodule_output), internals
-        elif dim == 3:
-            return self._to_3d(submodule_output), internals
-        elif dim == 4:
-            return self._to_4d(submodule_output), internals
-        else:
-            raise ValueError('Invalid dim: {}'.format(dim))
